updated define command
searching now works properly as far as i can tell. current code tries to find a "similar" match to user input in the list of keywords to check for typos, but if none are found it will search the definitions for a matching word, and kick back all definitions that contain the user input. if still no matches are found it will simply reply with "no matches found"